refactor(plugins): split manager.rs monolith into per-concern submodules
`crates/zeph-plugins/src/manager.rs` had grown to 4253 lines mixing install
lifecycle, remote registry interaction, security validation, and filesystem
state. Split it into `manager/{mod,install,registry,security,store,tests}.rs`:

- `install`  — add/remove/enable/disable and dependency guards
- `registry` — remote download, archive extraction, auto-update, ephemeral install
- `security` — URL/name/overlay validation, MCP allowlisting, archive safety, skill scanning
- `store`    — filesystem state and manifest reading
- `mod`      — public data types, `PluginManager` constructors, and re-exports

Public API is unchanged: `PluginManager`, `MAX_ARCHIVE_BYTES`,
`download_and_extract`, and `validate_url_scheme_ephemeral` remain reachable at
their existing paths, and the crate-internal `validate_plugin_name` /
`validate_overlay_keys` paths used by `overlay.rs` and `types.rs` are preserved
via re-exports. Cross-submodule helpers are bumped to `pub(crate)`; struct
fields stay private (submodules are descendants and retain field access).

Closes #4917
